+++ src/components/MewTokenContainer/MewTokenContainer.vue
@@ -2,7 +2,10 @@
   <!-- ===================================================================================== -->
   <!-- Mew Token Container -->
   <!-- ===================================================================================== -->
-  <div class="mew-token-container d-flex align-center justify-center">
+  <div
+    class="mew-token-container d-flex align-center justify-center"
+    :style="{ height: `${getSize}`, width: `${getSize}` }"
+  >
     <!-- ===================================================================================== -->
     <!-- Loading State -->
     <!-- ===================================================================================== -->
@@ -20,10 +23,9 @@
     <img
       v-if="!loading && img"
       :style="{ height: `${getSize}`, width: `${getSize}` }"
-      :src="img && !invalidImg ? img : ethTokenPlaceholder"
+      :src="img || ethTokenPlaceholder"
       :alt="name"
       loading="lazy"
-      @error="invalidImg = true"
     >
 
     <!-- ===================================================================================== -->
@@ -33,7 +35,7 @@
       v-if="!loading && !img"
       :class="[
         'd-flex align-center justify-center full-height textLight--text text-uppercase font-weight-medium',
-        getFontClass
+        getFontClass,
       ]"
     >
       {{ getPlaceholderText }}
@@ -52,39 +54,42 @@ export default {
      */
     loading: {
       type: Boolean,
-      default: false
+      default: false,
     },
     /**
      * Accepts small, medium or large sizes.
      */
     size: {
       type: String,
-      default: 'small'
+      default: 'small',
+    },
+    borderSize: {
+      type: String,
+      default: '',
     },
     /**
      * Token name. Used for placeholder if there is no icon img.
      */
     name: {
       type: String,
-      default: 'MEW'
+      default: 'MEW',
     },
     /**
      * Token Icon img src
      */
     img: {
       type: [String, Array],
-      default: ''
-    }
+      default: '',
+    },
   },
   data() {
     return {
-      invalidImg: false,
       ethTokenPlaceholder,
       sizeOptions: {
         small: 'small',
         medium: 'medium',
-        large: 'large'
-      }
+        large: 'large',
+      },
     };
   },
   computed: {
@@ -123,9 +128,13 @@ export default {
         return '32px';
       }
 
-      return '52px';
-    }
-  }
+      if (this.size.toLowerCase() === this.sizeOptions.large) {
+        return '52px';
+      }
+
+      return this.size;
+    },
+  },
 };
 </script>
 
@@ -134,7 +143,6 @@ export default {
   * has to be global styles to override vuetify
   */
 .mew-token-container {
-  height: 100%;
   overflow: hidden;
   background-color: var(--v-white-base);
   border-radius: 50%;