+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+### Release 1.3.3
+
+### devop
+
+- Add optional explorer for mew table [#301](https://github.com/MyEtherWallet/mew-components/pull/301)
+- remove select search border radius [#300](https://github.com/MyEtherWallet/mew-components/pull/300)
+- adjust button letter spacing [#299](https://github.com/MyEtherWallet/mew-components/pull/299)
+- update mew token container [#298](https://github.com/MyEtherWallet/mew-components/pull/298)
+- revert error for mewtokencontainer [#297](https://github.com/MyEtherWallet/mew-components/pull/297)
+
+### fix
+
+- mew-address-select showing wrong identicon bug [#296](https://github.com/MyEtherWallet/mew-components/pull/296)
 ### Release 1.1.0
 
 ### devop